+++ src/layouts/ExtensionWeb3Popup/containers/Web3DetectedContainer/Web3DetectedContainer.vue
@@ -2,7 +2,7 @@
   <div class="web-detected-container">
     <img src="~@/assets/images/icons/alien-spaceship.svg" />
     <p>
-      {{ parent.$t('mewCx.web3-detected') }}
+      {{ parent.$t('mewcx.web3-detected') }}
     </p>
   </div>
 </template>