+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 - Update translations[#1651](https://github.com/MyEtherWallet/MyEtherWallet/pull/1651)
 
+### Release v5.2.4
+
+### Devop
+
+- Update packages [#1669](https://github.com/MyEtherWallet/MyEtherWallet/pull/1669)
+- Make use of River nodes [#1568](https://github.com/MyEtherWallet/MyEtherWallet/pull/1568)
+
 ### Release v5.2.3-hotfix.2
 
 ### Devop
@@ -13,6 +20,7 @@
 - Fix toBuffer [#1658](https://github.com/MyEtherWallet/MyEtherWallet/pull/1658)
 - Fix dashboard dapps [#1650](https://github.com/MyEtherWallet/MyEtherWallet/pull/1650)
 - Fix MEWCX PromiEvent [#1646](https://github.com/MyEtherWallet/MyEtherWallet/pull/1646)
+- Fix MEWCX mobile ui [#1657](https://github.com/MyEtherWallet/MyEtherWallet/pull/1657)
 
 ### Release v5.2.3-hotfix.1